backend/requirements.txt
fastapi==0.115.5
email-validator==2.2.0
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.17
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
aiosqlite==0.20.0

# Existing project deps (from core/)
openpyxl==3.1.5
pandas==2.2.3
numpy==2.1.3
pdfplumber==0.11.4
openai==1.57.2
python-dotenv==1.0.1
python-magic==0.4.27
